requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
clamd==1.0.2
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.2.35
cryptography==39.0.0
defusedxml==0.7.1
Deprecated==1.2.13
deprecation==2.1.0
Django==3.2.18
django-celery-beat==2.5.0
django-celery-results==2.5.0
django-clamd==0.4.0
django-cors-headers==3.13.0
django-extensions==3.2.1
django-filter==22.1
django-hashid-field==3.3.7
django-js-asset==2.0.0
django-localflavor==3.1
django-model-utils==4.3.1
django-mptt==0.14.0
django-oauth-toolkit==2.2.0
django-signal-disabler==0.1.1
django-templated-mail==1.1.1
django-timezone-field==5.0
django-treebeard==4.7
django-vies==6.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-social-oauth2==1.2.1
ecdsa==0.18.0
et-xmlfile==1.1.0
geojson==3.0.1
gunicorn==20.1.0
hashids==1.3.1
idna==3.4
img2pdf==0.4.4
importlib-metadata==6.0.0
isodate==0.6.1
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.4.2
kombu==5.2.4
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.1
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.5.5.62
openpyxl==3.1.2
packaging==23.1
pandas==2.0.1
pdf2img==0.1.2
pikepdf==7.2.0
Pillow==9.3.0
placebo==0.9.0
platformdirs==2.6.2
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.6.0
pytesseract==0.3.10
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==0.21.0
python-jose==3.3.0
python-stdnum==1.18
python3-openid==3.2.0
pytz==2022.7
rauth==0.7.3
redis==4.5.4
requests==2.28.1
requests-auth==6.0.0
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
retrying==1.3.4
rsa==4.9
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
typing_extensions==4.4.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.2.0
wrapt==1.14.1
zeep==4.2.1
zipp==3.11.0
